bug(tests): moved korlin utils to new package
