+++ pom.xml
@@ -124,12 +124,6 @@
             <version>${mockito.version}</version>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.testcontainers</groupId>
-            <artifactId>junit-jupiter</artifactId>
-            <version>${testcontainers.version}</version>
-            <scope>test</scope>
-        </dependency>
         <!--Testcontainers-->
         <dependency>
             <groupId>org.testcontainers</groupId>
